pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drf-authentify"
version = "0.6.2"
description = "A simple authentication module for Django REST Framework"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
keywords = ["django", "django-rest-framework", "jwt", "authentication", "session-management", "security", "python", "drf"]

authors = [
    { name = "Gabriel Idenyi" }
]

requires-python = ">=3.9"

dependencies = [
    "Django>=4.2,<6.0",
    "djangorestframework>=3.14,<4.0",
    "PyJWT>=2.11,<3.0",
    "swapper>=1.3,<2.0",
    "typing_extensions; python_version < '3.11'",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Session",
]

[project.urls]
Homepage = "https://github.com/idenyigabriel/drf-sessions"
Documentation = "https://github.com/idenyigabriel/drf-sessions/blob/main/README.md"
Repository = "https://github.com/idenyigabriel/drf-sessions"
Issues = "https://github.com/idenyigabriel/drf-sessions/issues"
Changelog = "https://github.com/idenyigabriel/drf-sessions/blob/main/CHANGELOG.md"
